+++ sdk/@launchdarkly/observability-react-native/CHANGELOG.md
@@ -1,13 +1,24 @@
 # Changelog
 
+## 0.2.0
+
+### Minor Changes
+
+- ece4ed3: initial release of RN observability plugin
+
+### Patch Changes
+
+- cda2e4a: add tracingOrigins and urlBlocklist and improve span naming
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.0] - 2024-01-XX
+## [0.1.0] - 2025-07-08
 
 ### Added
+
 - Initial release of @launchdarkly/observability-react-native
 - OpenTelemetry instrumentation for React Native applications
 - Automatic tracing with fetch instrumentation